requirements.txt
numpy
python_dateutil
pytz
pandas
requests
logbook
anyjson
aiohttp
easytrader
easyquotation
arrow
tushare
xlrd
tables
json